requirements.txt
numpy
opencv-python-headless
Pillow
streamlit
streamlit-webrtc
av